Added controllers to the server folder and deleteRoutes.ts now uses deleteController.ts
