+++ server/controllers/deleteController.ts
@@ -1,18 +1,9 @@
-import ClientError from '../utils/client-error';
+const { fighters } = require('../model/user-database');
+import { Req, Res } from "../utils/types-routes";
+import ClientError from "../utils/client-error";
 import { client } from '../conn';
-import { Req, Res } from '../utils/types-routes';
 
-var express = require('express');
-const deleteRoutes = express.Router(); // Base route: 'api/delete'
-
-/**
- * Delete route that removes data from a table and id of your choice
- * @param { string } table // fighters, moves, throws, movements, or stats
- * @param { number } id   // fighterId, moveId, throwId, movementId, or statId
- * @return 204 status code
- */
-deleteRoutes.delete('/:table/:id', async (req: Req, res: Res, next: (param1: any) => any) => {
-
+async function deleteUnit(req: Req, res: Res, next: any) {
   const authHeader: string = req.headers['authorization'];
 
   try {
@@ -81,6 +72,8 @@ deleteRoutes.delete('/:table/:id', async (req: Req, res: Res, next: (param1: any
   } catch (e) {
     return next(e);
   }
-});
+}
 
-module.exports = deleteRoutes;
+module.exports = {
+  deleteUnit,
+}
