+++ server/controllers/getController.ts
@@ -1,0 +1,6 @@
+const { fighters } = require('../model/user-database');
+import { Req, Res } from "../utils/types-routes";
+
+async function getFighters(req: Req, res: Res, next: any) {
+
+}
