+++ server/controllers/getRoutes.ts
