+++ server/index.ts
@@ -30,11 +30,11 @@ app.use(staticMiddleware);
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 // All routes referenced here
-const getRoutes = require('./routes/get-routes');
-const postRoutes = require('./routes/post-routes');
-const updateRoutes = require('./routes/update-routes');
-const deleteRoutes = require('./routes/delete-routes');
-const registrationRoutes = require('./routes/registration-routes');
+const getRoutes = require('./controllers/getRoutes');
+const postRoutes = require('./controllers/postRoutes');
+const updateRoutes = require('./controllers/updateRoutes');
+const deleteRoutes = require('./controllers/deleteRoutes');
+const registrationRoutes = require('./controllers/registrationRoutes');
 app.use('/api/get', getRoutes);
 app.use('/api/add', postRoutes);
 app.use('/api/update', updateRoutes);
